+++ src/unix/index_unix.mli
@@ -18,6 +18,9 @@ all copies or substantial portions of the Software. *)
 module Make (K : Index.Key) (V : Index.Value) :
   Index.S with type key = K.t and type value = V.t
 
+module Syscalls = Syscalls
+(** Bindings to Unix system calls. *)
+
 (** These modules should not be used. They are exposed purely for testing
     purposes. *)
 module Private : sig