+++ CHANGES.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Added
+
+- Added `Index_unix.Syscalls`, a module exposing various Unix bindings for
+  interacting with file-systems.
+
 # 1.2.0 (2020-02-25)
 
 ## Added